Testing and API Specification added with other changes

+++ src/main/java/com/mediScreenDemoGraphics/controller/MediScreenController.java
@@ -13,8 +13,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
+
+import com.mediScreenDemoGraphics.exception.PatientIdNotFoundException;
 import com.mediScreenDemoGraphics.model.MediScreenPatient;
 import com.mediScreenDemoGraphics.service.MediScreenPatientServiceImpl;
+
 @RestController
 public class MediScreenController {
 	Logger logger = LoggerFactory.getLogger(MediScreenController.class);
@@ -24,37 +27,75 @@ public class MediScreenController {
 	MediScreenController(MediScreenPatientServiceImpl mediScreenPatientServiceImpl) {
 		this.mediScreenPatientServiceImpl = mediScreenPatientServiceImpl;
 	}
+	@GetMapping("/")
+	public ModelAndView landingPage() {
+		return new ModelAndView("home");
+	}
+	/**
+	 * 
+	 * @param patient
+	 * @return a form to add patient
+	 */
 
-	@GetMapping("patient/add")
-	public ModelAndView addPatient(@ModelAttribute("MediScreenPatient") MediScreenPatient patient) {
+	@GetMapping("/patient/add")
+	public ModelAndView showPatientForm(@ModelAttribute("mediScreenPatient") MediScreenPatient patient) {
 		logger.info("patient form");
 		return new ModelAndView("addPatient");
 	}
-
-	@PostMapping("patient/validate")
+	/**
+	 * 
+	 * @param patient
+	 * @param result
+	 * @param model
+	 * @return if all data are entered correctly a list of patient is display else a patient form is displayed
+	 */
+	@PostMapping("/patient/validate")
 	public ModelAndView validatePatient(@Valid MediScreenPatient patient, BindingResult result, Model model) {
 		if (!result.hasErrors()) {
 			mediScreenPatientServiceImpl.savePatient(patient);
 			logger.info("patient detials saved");
-			return new ModelAndView("redirect:/patientList");
+			return new ModelAndView("redirect:/patient/list");
 		} else {
 			return new ModelAndView("addPatient");
 		}
 	}
-	@GetMapping("patient/list")
+	/**
+	 * 
+	 * @param model
+	 * @return a list of all patients in DB is displayed in a UI Format
+	 */
+	@GetMapping("/patient/list")
 	public ModelAndView patientList(Model model) {
 		model.addAttribute("patients", mediScreenPatientServiceImpl.getAllPatients());
 		return new ModelAndView("patientList");
 	}
+/**
+ * 
+ * @param id
+ * @param model
+ * @return a form for modifying data stored in DB
+ * @throws PatientIdNotFoundException 
+ */
 	@GetMapping("/patient/update/{id}")
-	public ModelAndView showUpdateForm(@PathVariable("id") Integer id, Model model) {
+	public ModelAndView showUpdateForm(@PathVariable("id") Integer id, Model model) throws PatientIdNotFoundException {
 		MediScreenPatient patient = mediScreenPatientServiceImpl.getPatientById(id);
+		if(patient == null) {
+			throw new PatientIdNotFoundException(id);
+		}
 		model.addAttribute("patient", patient);
 		return new ModelAndView("updatePatient");
 	}
+	/**
+	 * 
+	 * @param id
+	 * @param patient
+	 * @param result
+	 * @param model
+	 * @return save modified patient information to db and displays a list of patients
+	 */
 	@PostMapping("/patient/update/{id}")
-	public ModelAndView updatePatientInfo(@PathVariable("id") Integer id, @Valid MediScreenPatient patient, BindingResult result,
-			Model model) {
+	public ModelAndView updatePatientInfo(@PathVariable("id") Integer id, @Valid MediScreenPatient patient,
+			BindingResult result, Model model) {
 		if (result.hasErrors()) {
 			return new ModelAndView("patient/update");
 		}
@@ -62,21 +103,35 @@ public class MediScreenController {
 		mediScreenPatientServiceImpl.savePatient(patient);
 		return new ModelAndView("redirect:/patient/list");
 	}
-	@GetMapping("/user/delete/{id}")
-	public ModelAndView deletePatient(@PathVariable("id") Integer id, Model model) {
+	/**
+	 * 
+	 * @param id
+	 * @param model
+	 * @return deletes a patient info by patient id
+	 * @throws PatientIdNotFoundException 
+	 */
+
+	@GetMapping("/patient/delete/{id}")
+	public ModelAndView deletePatient(@PathVariable("id") Integer id, Model model) throws PatientIdNotFoundException {
 		MediScreenPatient patient = mediScreenPatientServiceImpl.getPatientById(id);
+		if(patient==null) {
+			throw new PatientIdNotFoundException(id);
+		}
 		mediScreenPatientServiceImpl.deletePatient(patient);
 		return new ModelAndView("redirect:/patient/list");
 	}
-	@GetMapping("patient/info")
-	public ModelAndView getPatientInfoById(@PathVariable("id") Integer id, Model model) {
+	/**
+	 * 
+	 * @param id
+	 * @return patient info by id
+	 * @throws PatientIdNotFoundException 
+	 */
+	@GetMapping("/patient/info/{id}")
+	public MediScreenPatient getPatientInfoById(@PathVariable("id") Integer id) throws PatientIdNotFoundException {
 		MediScreenPatient patient = mediScreenPatientServiceImpl.getPatientById(id);
-		model.addAttribute("patientName",patient.getPatientName() );
-		model.addAttribute("patientFamilyName", patient.getFamilyName());
-		model.addAttribute("patientDateOfBirth", patient.getDateOfBirth());
-		model.addAttribute("patientHomeAddress", patient.getHomeAddress());
-		model.addAttribute("patientPhoneNumber", patient.getPhoneNumber());
-		model.addAttribute("patientSex", patient.getSex());
-		return new ModelAndView("patientInfo");
+		if (patient==null) {
+			throw new PatientIdNotFoundException(id);
+		}
+		return patient;
 	}
 }
